+++ ChessGameWebApp/Client/Components/AuthenticationComponent.razor
@@ -1,7 +1,7 @@
 ﻿@using AuthWebAPI
 @using AuthWebAPI.AuthWebAPI;
 @using Blazored.LocalStorage
-@using ChessGameWebApp.Client.Services
+@using AuthWebAPI.Services
 @using JwtToken
 @using Models
 @inherits AuthenticationComponentModel
