+++ ChessGameWebApp/Client/Components/ChatComponent.razor
@@ -1,5 +1,6 @@
-﻿@using ChessGameWebApp.Client.Pages
-@using ChessGameWebApp.Client.Services
+﻿@using AuthWebAPI.Models;
+@using ChessGameWebApp.Client.Pages
+@using AuthWebAPI.Services
 @using ChessGameWebApp.Shared
 @using Microsoft.AspNetCore.SignalR.Client;
 @using ChessGameWebApp.Client.Components
