+++ ChessGameWebApp/Client/Components/ChessBoardComponent.razor
@@ -1,4 +1,6 @@
-﻿@using ChessGame
+﻿@using AuthWebAPI.Models;
+@using AuthWebAPI.Services;
+@using ChessGame
 @using ChessGameWebApp.Client.Services;
 @inherits GameComponentModel
 @inject IGameHubService gameHub
