+++ ChessGameWebApp/Client/Components/GoGameComponent.razor.cs
@@ -1,5 +1,7 @@
 ﻿using AuthWebAPI;
 using AuthWebAPI.AuthWebAPI;
+using AuthWebAPI.Models;
+using AuthWebAPI.Services;
 using ChessGame;
 using ChessGameWebApp.Client.Services;
 using Microsoft.AspNetCore.Components;
