+++ ChessGameWebApp/Client/Components/InviteComponent.razor
@@ -1,4 +1,6 @@
-﻿@using ChessGameWebApp.Client.Services
+﻿@using AuthWebAPI.Models;
+@using AuthWebAPI.Services;
+@using ChessGameWebApp.Client.Services
 @inject SiteUserInfo User
 @implements IDisposable
 @inject IGameHubService gameHubService
