+++ ChessGameWebApp/Client/Components/LogOutComponent.razor
@@ -1,5 +1,5 @@
 ﻿@page "/logout"
-@using ChessGameWebApp.Client.Services
+@using AuthWebAPI.Services
 @inject IAuthService authService
 @inject NavigationManager navigationManager
