+++ ChessGameWebApp/Client/Components/StatusComponent.razor.cs
@@ -1,4 +1,5 @@
-﻿using ChessGame;
+﻿using AuthWebAPI.Services;
+using ChessGame;
 using ChessGameWebApp.Client.Services;
 using Microsoft.AspNetCore.Components;
