+++ ChessGameWebApp/Client/Components/TimerComponent.razor
@@ -1,5 +1,5 @@
 ﻿@using ChessGame
-@using ChessGameWebApp.Client.Models
+@using AuthWebAPI.Models
 @inject ChessBoard board
 @inject TimeUpdater updater
 @implements IChessObserver
