+++ ChessGameWebApp/Client/Pages/Game.razor
@@ -1,5 +1,6 @@
 ﻿@page "/game/{StartGame?}"
 @using AuthWebAPI.AuthWebAPI;
+@using AuthWebAPI.Models;
 @using ChessGameWebApp.Client.Components
 @using Microsoft.AspNetCore.SignalR.Client
 @using AuthWebAPI
