+++ ChessGameWebApp/Client/Pages/Index.razor
@@ -1,4 +1,5 @@
 ﻿@page "/"
+@using AuthWebAPI.Models;
 @using ChessGameWebApp.Client.Components
 @inject SiteUserInfo userInfo
