+++ ChessGameWebApp/Client/Program.cs
@@ -4,11 +4,13 @@ using ChessGameWebApp.Client.Services;
 using AuthWebAPI;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
-using ChessGameWebApp.Client.Models;
 using Blazored.Modal;
 using ChessGameWebApp.Shared;
 using ChessGameWebApp.Client.Services.Impl;
 using AuthWebAPI.AuthWebAPI;
+using AuthWebAPI.Services;
+using AuthWebAPI.Models;
+using AuthWebAPI.Services.Impl;
 
 var builder = WebAssemblyHostBuilder.CreateDefault(args);
 builder.RootComponents.Add<App>("#app");
