+++ ChessGameWebApp/Client/Shared/MainLayout.razor
@@ -1,4 +1,5 @@
-﻿@using Models
+﻿@using AuthWebAPI.Models;
+@using Models
 @using ChessGameWebApp.Client.Components
 @inherits LayoutComponentBase
 @inject SiteUserInfo User
