+++ ChessGameWebApp/Client/Shared/NavMenu.razor
@@ -1,4 +1,5 @@
-﻿@using ChessGame
+﻿@using AuthWebAPI.Models;
+@using ChessGame
 @inject SiteUserInfo user
 @implements IChessObserver
 @implements IDisposable
