+++ ChessWebAPI/Services/IClientGameService.cs
@@ -1,6 +1,6 @@
 ﻿using ChessGame;
 
-namespace ChessGameWebApp.Client.Services
+namespace AuthWebAPI.Services
 {
     public interface IClientGameService
     {
